Add channel argument to mailbox call function.
Although in lab1, we would only use channel 8 of mailbox.
But there is another channel 9 for handling requests from VC and responses from CPU.
So, it's better to use another argument to indicate which channel we would like to send our message.

Also, in this commit, I reduce the hang time to wait upon reboot.