Lab 1 C: Enable serial console locking
In contrast to early speculations, the implementation of the spinlock is
fine. The problem of spinlock acquisition hanging the processor arise
because the `ldaxrb` instruction cannot be used before the memory has
been set up properly and the data cache has been enabled. Therefore, we
opt to restrict the kernel to single-threaded execution and not lock the
serial console before the said conditions are met.

Reference:
https://stackoverflow.com/questions/68785276/bare-metal-spinlock-implementation-in-rust