Lab 4 C: Fix system lockup in `delay_ns`
The system lockup is caused by a race condition in the wfi loop. The
processor hangs forever under the scenario where the core timer IRQ
associated with the delay is taken after the flag is tested but before
the wfi instruction begins execution. This happened when running the
release kernel on a real machine.

Other minor fixes:

- Fix incorrect placement of volatile.