Lab 4 C: Fix lockup when using the console in ISR
When writing to the serial console while the write buffer is full, the
serial console enters a wfi loop, waiting for the ISR to clear the write
buffer. However, when the interrupt is masked (e.g. when executing
another ISR), the write buffer will never clear since the ISR
responsible for clearing the write buffer will not execute.
`_console_send_byte` now clears the write buffer by itself instead of
relying on the ISR to do so. A similar situation happens when reading
from the serial console in an ISR while the read buffer is empty.

The serial console also was susceptible to system lockup as a result of
race condition, the same situation as discussed in commit 03b8aca.