Lab 5 C: Remove many consistency checks
The consistency checks were meant to guarantee the correct operation of
the kernel by making sure that everything is loaded into the memory in a
sensible way (e.g. the memory region used by the kernel doesn't overlap
with other reserved regions). However, they make the code convoluted.
Given that the probability of something being incorrectly loaded is
exceedingly low in practice, it makes more sense not to include these
checks.

The quicksort implementation remains in the source tree, albeit becoming
unused.