Lab 5 C: Fix run queue corruption
`deliver_signal` depends on the invariant that, if the `is_waiting`
status bit is set on a thread, then that thread must be in a wait
queue. However, when waking up a thread from a wait queue, the
`is_waiting` status bit was not correctly cleared, causing
`deliver_signal` to corrupt the run queue.